frontend/requirements.txt
streamlit
requests
pillow
pandas
python-dotenv
altair
pydeck
beautifulsoup4
anthropic

